[Store] Action Creators now Receive Typed Reference to Store
- Fixes issue where it was necessary to cast the `store` argument in an action creator in order to be able to use `dispatch`
- Renames `Store` to `StoreType` and `MainStore` to `Store`
