+++ ReSwift/CoreTypes/Store.swift
@@ -1,216 +1,135 @@
 //
-//  Store.swift
-//  ReSwift
+//  MainStore.swift
+//  SwiftFlow
 //
-//  Created by Benjamin Encz on 11/28/15.
+//  Created by Benjamin Encz on 11/11/15.
 //  Copyright © 2015 DigiTales. All rights reserved.
 //
 
 import Foundation
 
 /**
-Defines the interface of Stores in Swift Flow. `MainStore` is the default implementation of this
-interaface. Applications have a single store that stores the entire application state.
-Stores receive actions and use reducers combined with these actions, to calculate state changes.
-Upon every state update a store informs all of its subscribers.
-*/
-public protocol Store {
-
-    typealias State: StateType
-
-    /// Initializes the store with a reducer and an intial state.
-    init(reducer: AnyReducer, state: State)
-
-    /// Initializes the store with a reducer, an initial state and a list of middleware.
-    /// Middleware is applied in the order in which it is passed into this constructor.
-    init(reducer: AnyReducer, state: State, middleware: [Middleware])
-
-    /// The current state stored in the store.
-    var state: State { get }
-
-    /**
-     The main dispatch function that is used by all convenience `dispatch` methods.
-     This dispatch function can be extended by providing middlewares.
-    */
-    var dispatchFunction: DispatchFunction! { get }
-
-    /**
-     Subscribes the provided subscriber to this store.
-     Subscribers will receive a call to `newState` whenever the
-     state in this store changes.
-
-     - parameter subscriber: Subscriber that will receive store updates
-     */
-    func subscribe(subscriber: AnyStoreSubscriber)
-
-    /**
-     Unsubscribes the provided subscriber. The subscriber will no longer
-     receive state updates from this store.
-
-     - parameter subscriber: Subscriber that will be unsubscribed
-    */
-    func unsubscribe(subscriber: AnyStoreSubscriber)
-
-    /**
-     Dispatches an action. This is the simplest way to modify the stores state.
-
-     Example of dispatching an action:
-
-     ```
-     store.dispatch( CounterAction.IncreaseCounter )
-     ```
-
-     - parameter action: The action that is being dispatched to the store
-     - returns: By default returns the dispatched action, but middlewares can change the
-     return type, e.g. to return promises
-     */
-    func dispatch(action: Action) -> Any
-
-    /**
-     Dispatches an action creator to the store. Action creators are functions that generate
-     actions. They are called by the store and receive the current state of the application
-     and a reference to the store as their input.
-
-     Based on that input the action creator can either return an action or not. Alternatively
-     the action creator can also perform an asynchronous operation and dispatch a new action
-     at the end of it.
-
-     Example of an action creator:
-
-     ```
-     func deleteNote(noteID: Int) -> ActionCreator {
-        return { state, store in
-            // only delete note if editing is enabled
-            if (state.editingEnabled == true) {
-                return NoteDataAction.DeleteNote(noteID)
-            } else {
-                return nil
-            }
-        }
-     }
-     ```
-
-     This action creator can then be dispatched as following:
-
-     ```
-     store.dispatch( noteActionCreatore.deleteNote(3) )
-     ```
-
-     - returns: By default returns the dispatched action, but middlewares can change the
-     return type, e.g. to return promises
-     */
-    func dispatch(actionCreator: ActionCreator) -> Any
-
-    /**
-     Dispatches an async action creator to the store. An async action creator generates an
-     action creator asynchronously.
-     */
-    func dispatch(asyncActionCreator: AsyncActionCreator)
-
-    /**
-     Dispatches an action and calls the callback as soon as the action has been processed.
-     You will receive the updated store state as part of this callback.
-
-     Example of dispatching an action and implementing a callback:
-
-     ```
-     store.dispatch( CounterAction.IncreaseCounter ) { state in
-        print("New state: \(state)")
-     }
-     ```
-
-     - parameter action: The action that is being dispatched to the store
-     - returns: By default returns the dispatched action, but middlewares can change the
-     return type, e.g. to return promises
-     */
-    func dispatch(action: Action, callback: DispatchCallback?) -> Any
-
-    /**
-     Dispatches an action creator to the store. Action creators are functions that generate
-     actions. They are called by the store and receive the current state of the application
-     and a reference to the store as their input.
-
-     Based on that input the action creator can either return an action or not. Alternatively
-     the action creator can also perform an asynchronous operation and dispatch a new action
-     at the end of it.
-
-     Example of an action creator:
-
-     ```
-     func deleteNote(noteID: Int) -> ActionCreator {
-        return { state, store in
-            // only delete note if editing is enabled
-            if (state.editingEnabled == true) {
-                return NoteDataAction.DeleteNote(noteID)
-            } else {
-                return nil
-            }
-        }
-     }
-     ```
-
-     This action creator can then be dispatched as following:
-
-     ```
-     store.dispatch( noteActionCreatore.deleteNote(3) )
-     ```
-
-     This overloaded version of `dispatch` will call the provided `callback` as soon as a new
-     state has been calculated based on the dispatch action.
-
-     - Note: If the ActionCreator does not dispatch an action, the callback block will never
-        be called
-
-     - returns: By default returns the dispatched action, but middlewares can change the
-     return type, e.g. to return promises
-     */
-    func dispatch(actionCreator: ActionCreator, callback: DispatchCallback?) -> Any
-
-    /**
-     Dispatches an async action creator to the store. An async action creator generates an
-     action creator asynchronously. Use this method if you want to wait for the state change
-     triggered by the asynchronously generated action creator.
-
-     This overloaded version of `dispatch` calls the provided `callback` as soon as the
-     asynchronoously dispatched action has caused a new state calculation.
-
-     - Note: If the ActionCreator does not dispatch an action, the callback block will never
-     be called
-     */
-    func dispatch(asyncActionCreator: AsyncActionCreator, callback: DispatchCallback?)
-}
-
-/**
- An optional callback that can be passed to the `dispatch` method.
- This callback will be called when the dispatched action triggers a new state calculation.
- This is useful when you need to wait on a state change, triggered by an action (e.g. wait on
- a successful login). However, you should try to use this callback very seldom as it
- deviates slighlty from the unidirectional data flow principal.
+ This class is the default implementation of the `Store` protocol. You will use this store in most
+ of your applications. You shouldn't need to implement your own store.
+ You initialize the store with a reducer and an initial application state. If your app has multiple
+ reducers you can combine them by initializng a `MainReducer` with all of your reducers as an
+ argument.
  */
-public typealias DispatchCallback = (StateType) -> Void
+public class Store<State: StateType>: StoreType {
 
-/**
-An ActionCreator is a function that, based on the received state argument, might or might not
-create an action.
+    // TODO: Setter should not be public; need way for store enhancers to modify appState anyway
 
-Example:
-
-```
-func deleteNote(noteID: Int) -> ActionCreator {
- return { state, store in
-    // only delete note if editing is enabled
-    if (state.editingEnabled == true) {
-        return NoteDataAction.DeleteNote(noteID)
-    } else {
-        return nil
+    /*private (set)*/ public var state: State! {
+        didSet {
+            subscribers.forEach { $0._newState(state) }
+        }
     }
- }
+
+    public var dispatchFunction: DispatchFunction!
+
+    private var reducer: AnyReducer
+    var subscribers: [AnyStoreSubscriber] = []
+    private var isDispatching = false
+
+    public required convenience init(reducer: AnyReducer, state: State?) {
+        self.init(reducer: reducer, state: state, middleware: [])
+    }
+
+    public required init(reducer: AnyReducer, state: State?, middleware: [Middleware]) {
+        self.reducer = reducer
+
+        // Wrap the dispatch function with all middlewares
+        self.dispatchFunction = middleware.reverse().reduce(self._defaultDispatch) {
+            dispatchFunction, middleware in
+            return middleware(self.dispatch, { self.state })(dispatchFunction)
+        }
+
+        if let state = state {
+            self.state = state
+        } else {
+            dispatch(SwiftFlowInit())
+        }
+    }
+
+    public func subscribe(subscriber: AnyStoreSubscriber) {
+        guard subscribers.indexOf({ $0 === subscriber }) == nil else {
+            print("Store subscriber is already added, ignoring.")
+            return
+        }
+
+        subscribers.append(subscriber)
+
+        if let state = self.state {
+            subscriber._newState(state)
+        }
+    }
+
+    public func unsubscribe(subscriber: AnyStoreSubscriber) {
+        let index = subscribers.indexOf { return $0 === subscriber }
+
+        if let index = index {
+            subscribers.removeAtIndex(index)
+        }
+    }
+
+    public func _defaultDispatch(action: Action) -> Any {
+        if isDispatching {
+            // Use Obj-C exception since throwing of exceptions can be verified through tests
+            NSException.raise("SwiftFlow:IllegalDispatchFromReducer", format: "Reducers may not " +
+                "dispatch actions.", arguments: getVaList(["nil"]))
+        }
+
+        isDispatching = true
+        let newState = self.reducer._handleAction(self.state, action: action) as! State
+        isDispatching = false
+
+        self.state = newState
+
+        return action
+    }
+
+    public func dispatch(action: Action) -> Any {
+        return dispatch(action, callback: nil)
+    }
+
+    public func dispatch(actionCreatorProvider: ActionCreator) -> Any {
+        return dispatch(actionCreatorProvider, callback: nil)
+    }
+
+    public func dispatch(asyncActionCreatorProvider: AsyncActionCreator) {
+        dispatch(asyncActionCreatorProvider, callback: nil)
+    }
+
+    public func dispatch(action: Action, callback: DispatchCallback?) -> Any {
+        let returnValue = self.dispatchFunction(action)
+        callback?(self.state)
+
+        return returnValue
+    }
+
+    public func dispatch(actionCreatorProvider: ActionCreator, callback: DispatchCallback?) -> Any {
+        let action = actionCreatorProvider(state: self.state, store: self)
+        if let action = action {
+            dispatch(action, callback: callback)
+        }
+
+        return action
+    }
+
+    public func dispatch(actionCreatorProvider: AsyncActionCreator, callback: DispatchCallback?) {
+        actionCreatorProvider(state: self.state, store: self) { actionProvider in
+            let action = actionProvider(state: self.state, store: self)
+            if let action = action {
+                self.dispatch(action, callback: callback)
+            }
+        }
+    }
+
+    public typealias DispatchCallback = (State) -> Void
+
+    public typealias ActionCreator = (state: State, store: Store) -> Action?
+
+    public typealias AsyncActionCreator = (state: State, store: Store,
+        actionCreatorCallback: ActionCreator -> Void) -> Void
+
 }
-```
-
-*/
-public typealias ActionCreator = (state: StateType, store: Store) -> Action?
-
-/// AsyncActionCreators allow the developer to wait for the completion of an async action.
-public typealias AsyncActionCreator = (state: StateType, store: Store,
-    actionCreatorCallback: ActionCreator -> Void) -> Void
